adoptedStylesheet test
